third_party/WebKit/LayoutTests/fast/forms/number/number-validity-badinput.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<style>
:invalid {
  background-color: #ff0000;
}
</style>
</head>
<body>
<div id=parent>
<input type=number id=number>
<input id=another>
</div>
<script>
description('A number input fields with a bad input string should make validity.badInput true and have :invalid style.');

function colorOf(el) {
    return document.defaultView.getComputedStyle(el, null).getPropertyValue('background-color');
}
var invalidStyleColor = 'rgb(255, 0, 0)';

var number = document.getElementById('number');
number.focus();
debug('Initial state. The elment has no value.');
shouldNotBe('colorOf(number)', 'invalidStyleColor');
shouldBeFalse('number.validity.badInput');

debug("Type '-'. The element becomes badInput.");
document.execCommand('InsertText', false, '-');
shouldBe('colorOf(number)', 'invalidStyleColor');
shouldBeFalse('number.validity.valid');
shouldBeTrue('number.validity.badInput');
shouldBeEqualToString('number.value', '');

debug("Type '1' additionally. The element becomes valid.");
document.execCommand('InsertText', false, '1');
shouldNotBe('colorOf(number)', 'invalidStyleColor');
shouldBeFalse('number.validity.badInput');
shouldBeEqualToString('number.value', '-1');

debug("Type '-' additionally. The element becomes badInput again.");
document.execCommand('InsertText', false, '-');
shouldBe('colorOf(number)', 'invalidStyleColor');
shouldBeTrue('number.validity.badInput');
shouldBeEqualToString('number.value', '');

debug("The element losts focus. The element state should not be changed.");
document.getElementById('another').focus();
shouldBe('colorOf(number)', 'invalidStyleColor');
shouldBeTrue('number.validity.badInput');
// Visible value is '-1-'.
number.focus();
document.execCommand('SelectAll');
shouldBeEqualToString('document.getSelection().toString()', '-1-');
shouldBeEqualToString('number.value', '');

debug("The element losts a renderer. The element state should not be changed.");
shouldBeTrue('number.style.display = "none"; number.validity.badInput');

number.style.display = 'inline-block';
number.focus();
debug('A bad input should be cleared by value="".');
shouldBeEqualToString('number.value = ""; document.execCommand("SelectAll"); document.getSelection().toString()', '');

debug('Setting a valud value via |value| IDL attribute should clear badInput.');
document.execCommand('InsertText', false, 'non-number');
shouldBeTrue('number.validity.badInput');
shouldBeFalse('number.checkValidity()');
number.value = '123';
shouldBeFalse('number.validity.badInput');
shouldBeTrue('number.checkValidity()');

number.value = '';
debug("' ' is not a badInput.");
document.execCommand('InsertText', false, ' ');
shouldBeFalse('number.validity.badInput');
shouldBeEqualToString('number.value', '');
debug("' +' is a badInput.");
document.execCommand('InsertText', false, '+');
shouldBeTrue('number.validity.badInput');
shouldBeEqualToString('number.value', '');
debug("' +1' is not a badInput.");
document.execCommand('InsertText', false, '1');
shouldBeFalse('number.validity.badInput');
shouldBeEqualToString('number.value', '1');
debug("' +1.' is not a badInput.");
document.execCommand('InsertText', false, '.');
shouldBeFalse('number.validity.badInput');
shouldBeEqualToString('number.value', '1');
debug("' +1. ' is not a badInput.");
document.execCommand('InsertText', false, ' ');
shouldBeFalse('number.validity.badInput');
shouldBeEqualToString('number.value', '1');

document.getElementById('parent').innerHTML = '';
</script>
</body>
</html>
